extract run lineage route
